Arduino controlled room heating system
The first thing that came to my mind when I started micro-controller
programming was to set an automatic room heating cum humidifier system.
Bangalore has a very confusing weather - especially for new-comers. The
temperature is generally moderate, but nights can sometimes become very
cold. However, on other days, the nights are warm and if you have a
heater on - it becomes uncomfortable. Also, since the weather is
generally moderate - automatic heating solutions are not available
commercially.

Hence, one of the first projects I made on an arduino was a temperature
and humidity controller.